Bug Fix for Creating New Issue
This fix the bug where we get `default index 1 out of range (1 - 0)` error when
we tried to create new Issue. This is because the Project object that we get
from list of Projects does not include Issue Types. Project's Issue Types
only available in individual Project API.

Refer to below issues for more troubleshooting details:

https://github.com/keepcosmos/terjira/issues/51